django_api/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
djangorestframework_simplejwt==4.6.0
PyJWT==2.8.0
pyotp==2.9.0
pytest==7.4.0
Requests==2.31.0
